Setup rewriting propoerly, and rewrite to the correct location
This assumes the s3 bucket is structured as follows:

```
/v2/<image>
  /manifests/<tag>
  /blobs/algorithm/hex
```

The content in /manifests/<tag> can actually be an empty file. They do
need to return the correct docker-content-digest header (can be done by
setting x-aws-meta-docker-content-digest with the correct digest) which
should be the manifest for the image.

The content in /blobs/algorithm/hex are the blobs produced by the build
process